client: GTM make sure all pages which use Layout are logged
