+++ client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "cva",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "A blog which uses Markdown and a Weaviate Knowledge graph to populate the content.",
   "keywords": [
     "Weaviate",
